requirements.txt
flask==3.1.0
fla=sk-cors=4.0.1
python-dotenv==1.0.1
gunicorn==23.0.0